feat: 채팅 컨텍스트 분리
